Added the result screen functionality
Two main components are shown here:

The first is the loading bar at the top, it leverages the linear
ease function in tailwind to smoothly flow up in tandem with the update
interval of the update function

NOTE: ANY UPDATE TO THE UPDATE INTERVAL FUNCTION SHOULD BE REFLECTED
IN THE DURATION OF THE COMPONENT, this could probably be done programmatically
if necessary

The second is the function for the pop in of result sentences, it makes
use of the alpine "nextTick" function to start hidden, and then subsequently show
the component on the next frame, to show the transition of the object in question.
In the case of the sentences a scale-in.