Added client manager & comms class as well as interface objects
Made a prototype server command and callback system that would
theoretically hookup to a gui renderer. We would still need to figure
out how to implement user input and user-driven events. For now the
client manager responds to certain server commands and updates state
and state data based on those events. Not very clean but it seems to pass some
of the custom tests I wrote with a nodejs test server.