Made it so that publicname only has to be given...
once. Before all payloads would need to include public id
and public name. This branch changes this so that it only has
to be given during either the lobby join or the user join call.
I will keep this in its own branch for now. To be merged if this
is the intended implementation.